requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
Flask==2.0.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.17.1
pytz==2021.3
requests==2.25.1
six==1.16.0
tmdbv3api==1.7.6
urllib3==1.26.7
Werkzeug==2.0.2
amqp==5.0.6
aniso8601==9.0.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
prompt-toolkit==3.0.22
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
